+++ lets/modules/disassemble/x86.py
@@ -1,0 +1,42 @@
+from lets.module import Module
+from lets.extensions.assembly import DisassemblyExtension
+
+# Imports required to execute this module
+import os, base64
+
+class X86(DisassemblyExtension, Module):
+    """
+    Disassemble bytes into x86 assembly code.
+    """
+
+    # A list of docker images required by the module.
+    images = [
+        "tools/capstone:latest"
+    ]
+
+    def do(self, data:bytes=None, options:dict=None) -> bytes:
+        """
+        Main functionality.
+
+        Module.do updates self.options with options.
+
+        DockerExtension.do prepares required docker images.
+
+        :param data: Data to be used by module, in bytes
+        :param options: Dict of options to be used by module
+        :return: Generator containing results of module execution, in bytes
+        """
+        super().do(data, options)
+
+        # Disassemble
+        yield self.disassemble(data, arch="X86", mode="32")
+
+    def test(self):
+        """
+        Perform unit tests to verify this module's functionality.
+        """
+        # Test production of results
+        sc = b"".join(self.do(b"\x90"))
+        self.assertGreater(len(sc),
+            0,
+            "NOP produced no results")